use a proper file for the type
