+++ packages/server/lib/controllers/connection.controller.ts
@@ -29,7 +29,7 @@ import {
 } from '@nangohq/shared';
 import { NANGO_ADMIN_UUID } from './account.controller.js';
 import { logContextGetter } from '@nangohq/logs';
-import type { RequestLocals } from '../utils/asyncWrapper.js';
+import type { RequestLocals } from '../utils/express.js';
 
 class ConnectionController {
     /**
