+++ packages/server/lib/controllers/environment.controller.ts
@@ -14,7 +14,7 @@ import {
     userService
 } from '@nangohq/shared';
 import { NANGO_ADMIN_UUID } from './account.controller.js';
-import type { RequestLocals } from '../utils/asyncWrapper.js';
+import type { RequestLocals } from '../utils/express.js';
 
 export interface GetMeta {
     environments: Pick<Environment, 'name'>[];
