+++ packages/server/lib/controllers/provider.controller.ts
@@ -1,6 +1,6 @@
 import type { NextFunction, Request, Response } from 'express';
 import { configService } from '@nangohq/shared';
-import type { RequestLocals } from '../utils/asyncWrapper';
+import type { RequestLocals } from '../utils/express.js';
 
 class ProviderController {
     /**
