+++ packages/server/lib/utils/express.ts
@@ -1,0 +1,8 @@
+import type { Account, Environment, User } from '@nangohq/shared';
+
+export interface RequestLocals {
+    authType?: 'secretKey' | 'publicKey' | 'basic' | 'adminKey' | 'none' | 'session';
+    user?: Pick<User, 'id' | 'email'>;
+    account?: Account;
+    environment?: Environment;
+}
